requirements.txt
Django==4.2.13
mysqlclient==2.2.7
PyMySQL==1.1.2
Pillow==12.0.0
djangorestframework==3.16.1
pytz==2025.2
sqlparse==0.5.3
asgiref==3.10.0
requests==2.32.5
firebase-admin==6.6.0
